requirements.txt
# Python 版本飞书机器人依赖包
# Python 3.10+ 推荐

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据验证和配置
pydantic==2.5.0
pydantic-settings==2.1.0

# 飞书官方 SDK
lark-oapi==1.4.23

# 配置文件解析
PyYAML==6.0.1
python-dotenv==1.0.0

# 异步文件操作
aiofiles==23.2.1

# 定时任务调度
APScheduler==3.10.4

# HTTP 客户端
httpx==0.25.2
requests==2.31.0

# 日志增强
python-json-logger==2.0.7

# 开发工具 (可选)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# 类型提示 (可选)
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
